+++ demo/src/server/middlewares/index.ts
@@ -3,3 +3,5 @@ export { authMiddleware } from './auth';
 export { tokenMiddleware } from './token';
 
 export { traceparentMiddleware } from './traceparent';
+
+export { serverTimingMiddleware } from './serverTiming';