+++ packages/web-sdk/src/instrumentations/performance/navigation.test.ts
@@ -78,11 +78,33 @@ describe('Navigation observer', () => {
     getNavigationTimings(mockPushEvent, ['']);
 
     expect(mockPushEvent).toHaveBeenCalledTimes(1);
-    expect(mockPushEvent).toHaveBeenCalledWith('faro.performance.navigation', {
-      ...createFaroResourceTiming(performanceNavigationEntry),
-      ...createFaroNavigationTiming(performanceNavigationEntry),
-      faroNavigationId: mockNavigationId,
-      faroPreviousNavigationId: 'unknown',
+    expect(mockPushEvent).toHaveBeenCalledWith(
+      'faro.performance.navigation',
+      {
+        ...createFaroResourceTiming(performanceNavigationEntry),
+        ...createFaroNavigationTiming(performanceNavigationEntry),
+        faroNavigationId: mockNavigationId,
+        faroPreviousNavigationId: 'unknown',
+      },
+      undefined,
+      {
+        spanContext: { traceId: '0af7651916cd43dd8448eb211c80319c', spanId: 'b7ad6b7169203331' },
+      }
+    );
+  });
+
+  it('Captures Server-Timings for w3c trace context', () => {
+    const mockPushEvent = jest.fn();
+    jest.spyOn(performanceUtilsModule, 'entryUrlIsIgnored').mockReturnValueOnce(false);
+
+    const mockNavigationId = '123';
+    jest.spyOn(faroCoreModule, 'genShortID').mockReturnValueOnce(mockNavigationId);
+
+    getNavigationTimings(mockPushEvent, ['']);
+
+    expect(mockPushEvent).toHaveBeenCalledTimes(1);
+    expect(mockPushEvent).toHaveBeenNthCalledWith(1, expect.anything(), expect.anything(), undefined, {
+      spanContext: { traceId: '0af7651916cd43dd8448eb211c80319c', spanId: 'b7ad6b7169203331' },
     });
   });
 
@@ -99,12 +121,19 @@ describe('Navigation observer', () => {
     getNavigationTimings(mockPushEvent, ['']);
 
     expect(mockPushEvent).toHaveBeenCalledTimes(1);
-    expect(mockPushEvent).toHaveBeenCalledWith('faro.performance.navigation', {
-      ...createFaroResourceTiming(performanceNavigationEntry),
-      ...createFaroNavigationTiming(performanceNavigationEntry),
-      faroNavigationId: mockNewNavigationId,
-      faroPreviousNavigationId: mockPreviousNavigationId,
-    });
+    expect(mockPushEvent).toHaveBeenCalledWith(
+      'faro.performance.navigation',
+      {
+        ...createFaroResourceTiming(performanceNavigationEntry),
+        ...createFaroNavigationTiming(performanceNavigationEntry),
+        faroNavigationId: mockNewNavigationId,
+        faroPreviousNavigationId: mockPreviousNavigationId,
+      },
+      undefined,
+      {
+        spanContext: { traceId: '0af7651916cd43dd8448eb211c80319c', spanId: 'b7ad6b7169203331' },
+      }
+    );
   });
 
   it('Stores navigationId in sessionStorage', () => {